Start refactoring.
