+++ derive/src/codegen.rs
@@ -1,4 +1,4 @@
-use crate::model::{ExtDefinition, ExternalFunc};
+use crate::parser::{ExtDefinition, ExternalFunc};
 use proc_macro2::{Ident, Literal, Span, TokenStream};
 use quote::{quote, quote_spanned, ToTokens};
